+++ yafsrc/YetAnotherForum.NET/Scripts/yaf.js
@@ -176,6 +176,18 @@ function mouseHover(cell, hover) {
 
 // Generic Functions
 jQuery(document).ready(function () {
+
+    // Cookie alert
+    if (!getCookie("YAF-AcceptCookies")) {
+        jQuery(".cookiealert").addClass("show");
+    }
+
+    $(".acceptcookies").click(function () {
+        setCookie("YAF-AcceptCookies", true, 60);
+        jQuery(".cookiealert").removeClass("show");
+    });
+
+
     
     /// <summary>
     /// Convert user posted image to modal images